backend/requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
pymongo